week05-object_detection/seminar/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
albumentations>=1.3.0
pycocotools>=2.0.6
scipy>=1.10.0
numpy>=1.24.0
pillow>=9.5.0
tensorboard>=2.13.0

